factor out shared Timer class
