+++ java/src/main/java/com/mlt/cli/Encode.java
@@ -1,4 +1,4 @@
-package com.mlt.cli;
+package com.mlt.tools;
 
 import org.apache.commons.cli.CommandLine;
 import org.apache.commons.cli.CommandLineParser;
@@ -16,6 +16,7 @@ import com.mlt.converter.mvt.ColumnMapping;
 import com.mlt.converter.mvt.MapboxVectorTile;
 import com.mlt.decoder.MltDecoder;
 import com.mlt.vector.FeatureTable;
+import com.mlt.tools.Timer;
 
 import java.nio.file.Files;
 import java.nio.file.Paths;
@@ -25,24 +26,6 @@ import java.util.List;
 import java.util.Optional;
 
 public class Encode {
-
-    public static class Timer {
-        private long startTime;
-
-        public Timer() {
-            startTime = System.nanoTime();
-        }
-
-        public void restart() {
-            startTime = System.nanoTime();
-        }
-
-        public void stop(String message) {
-            long endTime = System.nanoTime();
-            long elapsedTime = (endTime - startTime) / 1000000; // divide by 1000000 to get milliseconds
-            System.out.println("Time elapsed for " + message + ": " + elapsedTime + " milliseconds");
-        }
-    }
 
     public static void printMVT(MapboxVectorTile mvTile){
         var mvtLayers = mvTile.layers();
